requirements.txt
pyinstaller
GitPython
python-terraform